pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "polinews"
version = "0.3.0"
description = "Proof-of-Reading incentive layer — off-chain MVP"
requires-python = ">=3.11"
dependencies = [
    # Core
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.7",
    "python-dotenv>=1.0",
    # HTTP
    "httpx>=0.27",
    # Auth
    "python-jose[cryptography]>=3.3",
    # Scheduling
    "APScheduler>=3.10",
    # Rate-limiting
    "slowapi>=0.1.9",
    # Feed parsing
    "feedparser>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",  # TestClient
]

[project.scripts]
polinews-serve   = "polinews.main:serve"
polinews-collect = "polinews.main:collect"

[tool.setuptools.packages.find]
where = ["."]
include = ["polinews*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
